refactor: extract useExpertLogs and useExpertTutorial to resolve ESLint limits
